Week 2 Update
Week 2 Update

Changed
-/v1/characters/{character_id}/contracts/ updated (Added X-Pages +
volume changed from Float to Double)
-/v1/corporations/{corporation_id}/contracts/ updated (Added X-Pages +
volume changed from Float to Double)
-loyalty/stores/{corporation_id}/offers/ updated (iskCost changed from
BigDecimal to Long)

+++ src/main/java/net/troja/eve/esi/model/CorporationContractsResponse.java
@@ -206,7 +206,7 @@ public class CorporationContractsResponse implements Serializable {
     private Double buyout = null;
 
     @JsonProperty("volume")
-    private Float volume = null;
+    private Double volume = null;
 
     public CorporationContractsResponse contractId(Integer contractId) {
         this.contractId = contractId;
@@ -607,7 +607,7 @@ public class CorporationContractsResponse implements Serializable {
         this.buyout = buyout;
     }
 
-    public CorporationContractsResponse volume(Float volume) {
+    public CorporationContractsResponse volume(Double volume) {
         this.volume = volume;
         return this;
     }
@@ -618,11 +618,11 @@ public class CorporationContractsResponse implements Serializable {
      * @return volume
      **/
     @ApiModelProperty(example = "null", value = "Volume of items in the contract")
-    public Float getVolume() {
+    public Double getVolume() {
         return volume;
     }
 
-    public void setVolume(Float volume) {
+    public void setVolume(Double volume) {
         this.volume = volume;
     }
